refactor: split into header+cc
