+++ echion/mirrors.h
@@ -4,12 +4,18 @@
 
 #pragma once
 
+#include <memory>
+
 #define PY_SSIZE_T_CLEAN
+#define Py_BUILD_CORE
 #include <Python.h>
+#if PY_VERSION_HEX >= 0x030c0000
+// https://github.com/python/cpython/issues/108216#issuecomment-1696565797
+#undef _PyGC_FINALIZED
+#endif
+
 #include <dictobject.h>
 #include <setobject.h>
-
-#include <memory>
 
 #if PY_VERSION_HEX >= 0x030b0000
 #define Py_BUILD_CORE
@@ -68,12 +74,7 @@ public:
 class MirrorObject
 {
 public:
-    inline PyObject* reflect()
-    {
-        if (reflected == NULL)
-            throw MirrorError();
-        return reflected;
-    }
+    PyObject* reflect();
 
 protected:
     std::unique_ptr<char[]> data = nullptr;
@@ -86,62 +87,12 @@ class MirrorDict : public MirrorObject
 public:
     MirrorDict(PyObject*);
 
-    PyObject* get_item(PyObject* key)
-    {
-        return PyDict_GetItem(reflect(), key);
-    }
+    PyObject* get_item(PyObject* key);
 
 private:
     PyDictObject dict;
 };
 
-inline MirrorDict::MirrorDict(PyObject* dict_addr)
-{
-    if (copy_type(dict_addr, dict))
-        throw MirrorError();
-
-    PyDictKeysObject keys;
-    if (copy_type(dict.ma_keys, keys))
-        throw MirrorError();
-
-    // Compute the full dictionary data size
-#if PY_VERSION_HEX >= 0x030b0000
-    size_t entry_size =
-        keys.dk_kind == DICT_KEYS_UNICODE ? sizeof(PyDictUnicodeEntry) : sizeof(PyDictKeyEntry);
-    size_t keys_size = sizeof(PyDictKeysObject) + (1 << keys.dk_log2_index_bytes) +
-                       (keys.dk_nentries * entry_size);
-#else
-    size_t entry_size = sizeof(PyDictKeyEntry);
-    size_t keys_size = sizeof(PyDictKeysObject) + (keys.dk_size * sizeof(Py_ssize_t)) +
-                       (keys.dk_nentries * entry_size);
-#endif
-    size_t values_size = dict.ma_values != NULL ? keys.dk_nentries * sizeof(PyObject*) : 0;
-
-    // Allocate the buffer
-    ssize_t data_size = keys_size + (keys.dk_nentries * entry_size) + values_size;
-    if (data_size < 0 || data_size > (1 << 20))
-        throw MirrorError();
-
-    data = std::make_unique<char[]>(data_size);
-
-    // Copy the key data and update the pointer
-    if (copy_generic(dict.ma_keys, data.get(), keys_size))
-        throw MirrorError();
-
-    dict.ma_keys = (PyDictKeysObject*)data.get();
-
-    if (dict.ma_values != NULL)
-    {
-        // Copy the value data and update the pointer
-        char* values_addr = data.get() + keys_size;
-        if (copy_generic(dict.ma_values, keys_size, values_size))
-            throw MirrorError();
-
-        dict.ma_values = (PyDictValues*)values_addr;
-    }
-
-    reflected = (PyObject*)&dict;
-}
 
 // ----------------------------------------------------------------------------
 class MirrorSet : public MirrorObject
@@ -154,39 +105,3 @@ private:
     size_t size;
     PySetObject set;
 };
-
-inline MirrorSet::MirrorSet(PyObject* set_addr)
-{
-    if (copy_type(set_addr, set))
-        throw MirrorError();
-
-    size = set.mask + 1;
-    ssize_t table_size = size * sizeof(setentry);
-    if (table_size < 0 || table_size > (1 << 20))
-        throw MirrorError();
-
-    data = std::make_unique<char[]>(table_size);
-    if (copy_generic(set.table, data.get(), table_size))
-        throw MirrorError();
-
-    set.table = (setentry*)data.get();
-
-    reflected = (PyObject*)&set;
-}
-
-inline std::unordered_set<PyObject*> MirrorSet::as_unordered_set()
-{
-    if (data == nullptr)
-        throw MirrorError();
-
-    std::unordered_set<PyObject*> uset;
-
-    for (size_t i = 0; i < size; i++)
-    {
-        auto entry = set.table[i];
-        if (entry.key != NULL)
-            uset.insert(entry.key);
-    }
-
-    return uset;
-}
