+++ echion/stacks.h
@@ -6,11 +6,14 @@
 
 #define PY_SSIZE_T_CLEAN
 #include <Python.h>
+#if PY_VERSION_HEX >= 0x030c0000
+// https://github.com/python/cpython/issues/108216#issuecomment-1696565797
+#undef _PyGC_FINALIZED
+#endif
 
 #include <deque>
 #include <mutex>
 #include <unordered_map>
-#include <unordered_set>
 
 #ifndef UNWIND_NATIVE_DISABLE
 #define UNW_LOCAL_ONLY
@@ -30,50 +33,17 @@ public:
     using Key = Frame::Key;
 
     // ------------------------------------------------------------------------
-    Key key()
-    {
-        Key h = 0;
-
-        for (auto it = this->begin(); it != this->end(); ++it)
-            h = rotl(h) ^ (*it).get().cache_key;
-
-        return h;
-    }
+    Key key();
 
     // ------------------------------------------------------------------------
-    void render()
-    {
-        for (auto it = this->rbegin(); it != this->rend(); ++it)
-        {
-#if PY_VERSION_HEX >= 0x030c0000
-            if ((*it).get().is_entry)
-                // This is a shim frame so we skip it.
-                continue;
-#endif
-            Renderer::get().render_frame((*it).get());
-        }
-    }
+    void render();
 
     // ------------------------------------------------------------------------
-    void render_where()
-    {
-        for (auto it = this->rbegin(); it != this->rend(); ++it)
-        {
-#if PY_VERSION_HEX >= 0x030c0000
-            if ((*it).get().is_entry)
-                // This is a shim frame so we skip it.
-                continue;
-#endif
-            WhereRenderer::get().render_frame((*it).get());
-        }
-    }
+    void render_where();
 
 private:
     // ------------------------------------------------------------------------
-    static inline Frame::Key rotl(Key key)
-    {
-        return (key << 1) | (key >> (CHAR_BIT * sizeof(key) - 1));
-    }
+    static Frame::Key rotl(Key key);
 };
 
 // ----------------------------------------------------------------------------
@@ -84,240 +54,29 @@ inline FrameStack interleaved_stack;
 
 // ----------------------------------------------------------------------------
 #ifndef UNWIND_NATIVE_DISABLE
-inline void unwind_native_stack()
-{
-    unw_cursor_t cursor;
-    unw_context_t context;
-
-    unw_getcontext(&context);
-    unw_init_local(&cursor, &context);
-
-    native_stack.clear();
-
-    while (unw_step(&cursor) > 0 && native_stack.size() < max_frames)
-    {
-        try
-        {
-            native_stack.push_back(Frame::get(cursor));
-        }
-        catch (Frame::Error&)
-        {
-            break;
-        }
-    }
-}
+void unwind_native_stack();
 #endif  // UNWIND_NATIVE_DISABLE
 
 // ----------------------------------------------------------------------------
-static size_t unwind_frame(PyObject* frame_addr, FrameStack& stack)
-{
-    std::unordered_set<PyObject*> seen_frames;  // Used to detect cycles in the stack
-    int count = 0;
-
-    PyObject* current_frame_addr = frame_addr;
-    while (current_frame_addr != NULL && stack.size() < max_frames)
-    {
-        if (seen_frames.find(current_frame_addr) != seen_frames.end())
-            break;
-
-        seen_frames.insert(current_frame_addr);
-
-        try
-        {
-#if PY_VERSION_HEX >= 0x030b0000
-            Frame& frame =
-                Frame::read(reinterpret_cast<_PyInterpreterFrame*>(current_frame_addr),
-                            reinterpret_cast<_PyInterpreterFrame**>(&current_frame_addr));
-#else
-            Frame& frame = Frame::read(current_frame_addr, &current_frame_addr);
-#endif
-            stack.push_back(frame);
-        }
-        catch (Frame::Error& e)
-        {
-            break;
-        }
-
-        count++;
-    }
-
-    return count;
-}
+size_t unwind_frame(PyObject* frame_addr, FrameStack& stack);
 
 // ----------------------------------------------------------------------------
-static size_t unwind_frame_unsafe(PyObject* frame, FrameStack& stack)
-{
-    std::unordered_set<PyObject*> seen_frames;  // Used to detect cycles in the stack
-    int count = 0;
-
-    PyObject* current_frame = frame;
-    while (current_frame != NULL && stack.size() < max_frames)
-    {
-        if (seen_frames.find(current_frame) != seen_frames.end())
-            break;
-
-#if PY_VERSION_HEX >= 0x030d0000
-        // See the comment in unwind_frame()
-        while (current_frame != NULL)
-        {
-            if (((_PyInterpreterFrame*)current_frame)->f_executable->ob_type == &PyCode_Type)
-            {
-                break;
-            }
-            current_frame = (PyObject*)((_PyInterpreterFrame*)current_frame)->previous;
-        }
-
-        if (current_frame == NULL)
-        {
-            break;
-        }
-#endif  // PY_VERSION_HEX >= 0x030d0000
-        count++;
-
-        seen_frames.insert(current_frame);
-
-        stack.push_back(Frame::get(current_frame));
-
-#if PY_VERSION_HEX >= 0x030b0000
-        current_frame = (PyObject*)((_PyInterpreterFrame*)current_frame)->previous;
-#else
-        current_frame = (PyObject*)((PyFrameObject*)current_frame)->f_back;
-#endif
-    }
-
-    return count;
-}
+size_t unwind_frame_unsafe(PyObject* frame, FrameStack& stack);
 
 // ----------------------------------------------------------------------------
-static void unwind_python_stack(PyThreadState* tstate, FrameStack& stack)
-{
-    stack.clear();
-#if PY_VERSION_HEX >= 0x030b0000
-    try
-    {
-        if (stack_chunk == nullptr)
-        {
-            stack_chunk = std::make_unique<StackChunk>();
-        }
-        stack_chunk->update((_PyStackChunk*)tstate->datastack_chunk);
-    }
-    catch (StackChunkError& e)
-    {
-        stack_chunk = nullptr;
-    }
-#endif
-
-#if PY_VERSION_HEX >= 0x030d0000
-    PyObject* frame_addr = (PyObject*)tstate->current_frame;
-#elif PY_VERSION_HEX >= 0x030b0000
-    _PyCFrame cframe;
-    _PyCFrame* cframe_addr = tstate->cframe;
-    if (copy_type(cframe_addr, cframe))
-        // TODO: Invalid frame
-        return;
-
-    PyObject* frame_addr = (PyObject*)cframe.current_frame;
-#else  // Python < 3.11
-    PyObject* frame_addr = (PyObject*)tstate->frame;
-#endif
-    unwind_frame(frame_addr, stack);
-}
+void unwind_python_stack(PyThreadState* tstate, FrameStack& stack);
 
 // ----------------------------------------------------------------------------
-static void unwind_python_stack_unsafe(PyThreadState* tstate, FrameStack& stack)
-{
-    stack.clear();
-#if PY_VERSION_HEX >= 0x030b0000
-    try
-    {
-        if (stack_chunk == nullptr)
-        {
-            stack_chunk = std::make_unique<StackChunk>();
-        }
-        stack_chunk->update((_PyStackChunk*)tstate->datastack_chunk);
-    }
-    catch (StackChunkError& e)
-    {
-        stack_chunk = nullptr;
-    }
-#endif
-
-#if PY_VERSION_HEX >= 0x030d0000
-    PyObject* frame_addr = (PyObject*)tstate->current_frame;
-#elif PY_VERSION_HEX >= 0x030b0000
-    PyObject* frame_addr = (PyObject*)tstate->cframe->current_frame;
-#else  // Python < 3.11
-    PyObject* frame_addr = (PyObject*)tstate->frame;
-#endif
-    unwind_frame_unsafe(frame_addr, stack);
-}
+void unwind_python_stack_unsafe(PyThreadState* tstate, FrameStack& stack);
 
 // ----------------------------------------------------------------------------
-static void unwind_python_stack(PyThreadState* tstate)
-{
-    unwind_python_stack(tstate, python_stack);
-}
+void unwind_python_stack(PyThreadState* tstate);
 
 // ----------------------------------------------------------------------------
-static void interleave_stacks(FrameStack& python_stack)
-{
-    interleaved_stack.clear();
-
-    auto p = python_stack.rbegin();
-    // The last two frames are usually the signal trampoline and the signal
-    // handler. We skip them.
-    for (auto n = native_stack.rbegin(); n != native_stack.rend() - 2; ++n)
-    {
-        auto native_frame = *n;
-
-        if (string_table.lookup(native_frame.get().name).find("PyEval_EvalFrameDefault") !=
-            std::string::npos)
-        {
-            if (p == python_stack.rend())
-            {
-                // We expected a Python frame but we found none, so we report
-                // the native frame instead.
-                std::cerr << "Expected Python frame(s), found none!" << std::endl;
-                interleaved_stack.push_front(native_frame);
-            }
-            else
-            {
-                // We skip the PyEval_EvalFrameDefault frame because it is the
-                // function that calls the Python code.
-#if PY_VERSION_HEX >= 0x030b0000
-                int cframe_count = 0;
-                while (p != python_stack.rend())
-                {
-                    // The Python stack will start with an entry frame at the top.
-                    // We stop popping at the next entry frame.
-                    cframe_count += (*p).get().is_entry;
-                    if (cframe_count >= 2)
-                        break;
-
-                    interleaved_stack.push_front(*p++);
-                }
-#else
-                interleaved_stack.push_front(*p++);
-#endif
-            }
-        }
-        else
-            interleaved_stack.push_front(native_frame);
-    }
-
-    if (p != python_stack.rend())
-    {
-        std::cerr << "Python stack not empty after interleaving!" << std::endl;
-        while (p != python_stack.rend())
-            interleaved_stack.push_front(*p++);
-    }
-}
+void interleave_stacks(FrameStack& python_stack);
 
 // ----------------------------------------------------------------------------
-static void interleave_stacks()
-{
-    interleave_stacks(python_stack);
-}
+void interleave_stacks();
 
 // ----------------------------------------------------------------------------
 class StackInfo
@@ -327,7 +86,7 @@ public:
     bool on_cpu;
     FrameStack stack;
 
-    StackInfo(StringTable::Key task_name, bool on_cpu) : task_name(task_name), on_cpu(on_cpu) {}
+    StackInfo(StringTable::Key task_name, bool on_cpu);
 };
 
 // ----------------------------------------------------------------------------
@@ -337,40 +96,13 @@ class StackTable
 {
 public:
     // ------------------------------------------------------------------------
-    FrameStack::Key inline store(FrameStack::Ptr stack)
-    {
-        std::lock_guard<std::mutex> lock(this->lock);
-
-        auto stack_key = stack->key();
-
-        auto stack_entry = table.find(stack_key);
-        if (stack_entry == table.end())
-        {
-            table.emplace(stack_key, std::move(stack));
-        }
-        else
-        {
-            // TODO: Check for collisions.
-        }
-
-        return stack_key;
-    }
+    FrameStack::Key store(FrameStack::Ptr stack);
 
     // ------------------------------------------------------------------------
-    FrameStack& retrieve(FrameStack::Key stack_key)
-    {
-        std::lock_guard<std::mutex> lock(this->lock);
-
-        return *table.find(stack_key)->second;
-    }
+    FrameStack& retrieve(FrameStack::Key stack_key);
 
     // ------------------------------------------------------------------------
-    void clear()
-    {
-        std::lock_guard<std::mutex> lock(this->lock);
-
-        table.clear();
-    }
+    void clear();
 
 private:
     std::unordered_map<FrameStack::Key, std::unique_ptr<FrameStack>> table;
