+++ echion/timing.h
@@ -4,14 +4,13 @@
 
 #pragma once
 
-#if defined PL_LINUX
-#include <time.h>
-#elif defined PL_DARWIN
+#if defined PL_DARWIN
 #include <mach/clock.h>
 #include <mach/mach.h>
 
 inline clock_serv_t cclock;
 #endif
+
 
 typedef unsigned long microsecond_t;
 
@@ -21,16 +20,4 @@ inline microsecond_t last_time = 0;
 #define TV_TO_MICROSECOND(tv) ((tv).seconds * 1e6 + (tv).microseconds)
 
 // ----------------------------------------------------------------------------
-static microsecond_t gettime()
-{
-#if defined PL_LINUX
-    struct timespec ts;
-    if (clock_gettime(CLOCK_BOOTTIME, &ts))
-        return 0;
-    return TS_TO_MICROSECOND(ts);
-#elif defined PL_DARWIN
-    mach_timespec_t ts;
-    clock_get_time(cclock, &ts);
-    return TS_TO_MICROSECOND(ts);
-#endif
-}
+microsecond_t gettime();
