+++ setup.py
@@ -36,7 +36,25 @@ if DISABLE_NATIVE:
 
 echionmodule = Extension(
     "echion.core",
-    sources=["echion/coremodule.cc", "echion/frame.cc", "echion/render.cc"],
+    sources=[
+        "echion/coremodule.cc",
+        "echion/frame.cc",
+        "echion/render.cc",
+        "echion/config.cc",
+        "echion/greenlets.cc",
+        "echion/interp.cc",
+        "echion/long.cc",
+        "echion/memory.cc",
+        "echion/mirrors.cc",
+        "echion/signals.cc",
+        "echion/stack_chunk.cc",
+        "echion/stacks.cc",
+        "echion/strings.cc",
+        "echion/tasks.cc",
+        "echion/threads.cc",
+        "echion/timing.cc",
+        "echion/vm.cc",
+    ],
     include_dirs=["."],
     define_macros=[(f"PL_{PLATFORM.upper()}", None)],
     extra_compile_args=["-std=c++17", "-Wall", "-Wextra"] + CFLAGS + COLORS,
